refactor: centralize discord attachment helpers
